requirements.txt
# ── Core Framework ──
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# ── LLM & LangChain ──
langchain==0.3.7
langchain-groq==0.2.1
langchain-core>=0.3.17,<0.4.0
langchain-community==0.3.7

# ── Database ──
psycopg2-binary==2.9.9

# ── Utilities ──
httpx==0.27.2
tiktoken==0.8.0
